Update VMC example to use convenience method
The 2.4 NSX-T SDK now has a convenience method for authenticating
to VMC. This method only requires that the caller provide the
VMC organization ID, SDDC ID, and refresh token. The convenience
method automatically obtains an API session token and computes
the correct URL to access the NSX-T API. It is no longer necessary
to add NAT rules to allow access to the NSX-T Manager. The VMC
reverse proxy forwards the request.